+++ CHANGELOG.md
@@ -4,13 +4,17 @@
 
 **Features**:
 
+- Relay is now able to ingest pre-aggregated sessions, which will make it possible to efficiently handle applications that produce thousands of sessions per second. ([#815](https://github.com/getsentry/relay/pull/815))
 - Add dynamic sampling for transactions. ([#835](https://github.com/getsentry/relay/pull/835))
 
 **Bug Fixes**:
 
+- Allow params in content-type for security requests to support content types like `"application/expect-ct-report+json; charset=utf-8"`. ([#844](https://github.com/getsentry/relay/pull/844))
 
-**Internal**:
 
+## 20.11.1
+
+- No documented changes.
 
 ## 20.11.0